Extract Prolongation behaviour from model
This behaviour now contained in ProlongationManager class, which
takes control on possible operation with Prolongation model
in views.
