+++ openprocurement/auctions/core/plugins/contracting/v3/views/contract.py
@@ -10,7 +10,8 @@ from openprocurement.auctions.core.utils import (
     save_auction,
     opresource,
 )
-from openprocurement.auctions.core.plugins.contracting.v3.utils import (
+from openprocurement.auctions.core.plugins.\
+        contracting.v3.utils.common import (
     check_auction_status,
 )
 from openprocurement.auctions.core.validation import (
